% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{bf.create.ts}
\alias{bf.create.ts}
\title{#' bf.create.ts
#'
#' @param resp
setGeneric("bf.create.ts", function(resp) {NULL})}
\usage{
bf.create.ts(resp)
}
\arguments{
\item{resp}{response object from server}
}
\description{
#' bf.create.ts
#'
#' @param resp
setGeneric("bf.create.ts", function(resp) {NULL})
}
