% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/comms.R, R/core.R
\docType{class}
\name{BFClient-class}
\alias{BFClient-class}
\alias{bf.get,BFClient,character,list-method}
\alias{bf.post,BFClient,character,character-method}
\alias{bf.delete,BFClient,character,character-method}
\alias{bf.put,BFClient,character,character-method}
\alias{initialize,BFClient-method}
\alias{show,BFClient-method}
\alias{bf_datasets,BFClient-method}
\alias{contents,BFClient,BFDataset-method}
\alias{contents.id,BFClient,character-method}
\title{Class BFClient.}
\usage{
\S4method{bf.get}{BFClient,character,list}(client, path, params)

\S4method{bf.post}{BFClient,character,character}(client, path, params)

\S4method{bf.delete}{BFClient,character,character}(client, path, params)

\S4method{bf.put}{BFClient,character,character}(client, path, params)

\S4method{initialize}{BFClient}(.Object, ...)

\S4method{show}{BFClient}(object)

\S4method{bf_datasets}{BFClient}(object)

\S4method{contents}{BFClient,BFDataset}(client, target)

\S4method{contents.id}{BFClient,character}(client, id)
}
\arguments{
\item{client}{BFClient}

\item{path}{"character"}

\item{params}{"list"}

\item{object}{BFClient object}

\item{target}{"BFDataset" Id of a Dataset, or Collection}

\item{client}{BFClient}

\item{path}{""}

\item{params}{""}

\item{client}{BFClient}

\item{path}{""}

\item{params}{""}

\item{client}{BFClient}

\item{path}{""}

\item{params}{""}

\item{object}{BFClient object}

\item{client}{BFClient client}
}
\value{
BFClient object
}
\description{
This is the main interface class for the Blackfynn R
client. It handles the session and should be passed to
all method calls.
}
\section{Methods (by generic)}{
\itemize{
\item \code{bf.get}: 

\item \code{bf.post}: 

\item \code{bf.delete}: 

\item \code{bf.put}: 

\item \code{initialize}: Initializes the BFClient object. It opens
a Blackfynn session and stores the API token for the
session.

\item \code{show}: Shows a representation of a Blackfynn Client object.

\item \code{bf_datasets}: Returns a list of Datasets for the organization.
This is mroe inof

\item \code{contents}: Lists the content of a dataset

\item \code{contents.id}: Get the contents of any object on Blackfynn

    @param client BFClient client
    @param id "Character" ID of an object in a dataset
    @export
}}

\section{Slots}{

\describe{
\item{\code{scope}}{The scope in the configuration file}

\item{\code{org}}{The organization for the current session}
}}

