% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{contents}
\alias{contents}
\title{contents}
\usage{
contents(client, target)
}
\arguments{
\item{target}{Object for which contents should be returned}

\item{object}{BFCLient object}
}
\description{
contents
}
