% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/core.R
\docType{class}
\name{BFClient-class}
\alias{BFClient-class}
\alias{initialize,BFClient-method}
\alias{show,BFClient-method}
\alias{bf_datasets,BFClient-method}
\title{Class BFClient.}
\usage{
\S4method{initialize}{BFClient}(.Object, ...)

\S4method{show}{BFClient}(object)

\S4method{bf_datasets}{BFClient}(object)
}
\arguments{
\item{object}{BFClient object}
}
\value{
BFClient object
}
\description{
This is the main interface class for the Blackfynn R
client. It handles the session and should be passed to
all method calls.

This method renders objects of the BFClient class
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initializes the BFClient object. It opens
a Blackfynn session and stores the API token for the
session.

\item \code{show}: Shows a representation of a Blackfynn Client object.

\item \code{bf_datasets}: Returns a list of Datasets for the organization.
This is mroe inof
}}

\section{Slots}{

\describe{
\item{\code{scope}}{The scope in the configuration file}

\item{\code{org}}{The organization for the current session}
}}

